requirements.txt
# Core dependencies
crewai>=0.28.0
python-dotenv>=1.0.0

# LLM and API integrations
ibm-watson-machine-learning>=1.0.312
langchain>=0.0.313
langchain-community>=0.0.13
tavily-python>=0.2.6

# Web scraping tools
beautifulsoup4>=4.12.2
requests>=2.31.0
lxml>=4.9.3

# Utilities
pyyaml>=6.0.1
jsonschema>=4.19.1
pydantic>=2.4.2

# PDF generation
fpdf>=1.7.2

# Web application
streamlit>=1.22.0
pandas>=1.5.3
plotly>=5.14.1

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Type checking
mypy>=1.5.1
types-requests>=2.31.0.2
types-PyYAML>=6.0.12.11